Refactor the code and add milestone update validation

+++ src/controllers/propertyControllers.js
@@ -2,7 +2,7 @@ import Property from '../models/propertyModel.js';
 import Milestone from "../models/milestoneModel.js";
 import Wallet from '../models/wallet.js'
 import Transaction from '../models/transaction.js'
-
+import { isValidObjectId } from "mongoose";
 import { validateProperty, validateStatus } from '../utils/validation/propertyValidation.js'
 
 export const setProperty = async (req, res) => {
@@ -18,6 +18,7 @@ export const setProperty = async (req, res) => {
             return;
         }
 
+        // check partner percentage should be 100
         let PartnerPercentage = checkPartnerPercentage(propertyData.partner);
 
         if (!PartnerPercentage) {
@@ -28,6 +29,7 @@ export const setProperty = async (req, res) => {
 
         propertyData.created_by = userId;
 
+        // check property already exist or not
         const existingProperty = await Property.findOne({ property_name: propertyData.property_name, created_by: userId });
         if (existingProperty) {
             res.status(400).json({ status: 400, message: "Property already exist", data: null });
@@ -35,36 +37,34 @@ export const setProperty = async (req, res) => {
         }
 
         const property = new Property(propertyData);
-
         let saveProperty = await property.save();
 
-        if (saveProperty.status === "purchase") {
+        // check property saved or not 
+        if (saveProperty.length === 0 || !saveProperty) {
+            console.log(error);
+            res.status(500).json({ status: 500, message: "Error in adding property", data: null });
+            return;
+        }
+
+        // check property status is purchase or not if purchase then create milestone, transaction and update wallet
+        if (saveProperty?.status === "purchase") {
+
             let milestoneData = await createMilestone(property, userId);
-            // console.log("milestoneData->", milestoneData);
-
             let transactionData = await createTransaction(property, userId);
-            // console.log("transactionData->", transactionData);
-
             let walletData = await updateWallet(property, userId);
-            // console.log("walletData->", walletData);
+
+            // check wallet balance
             if (!walletData) {
                 res.status(400).json({ status: 400, message: "You don't have sufficient amount to purchase", data: null });
                 return;
             }
-
             res.status(201).json({ status: 201, message: "Property added successfully", data: property });
-            return;
-        }
-
-        if (saveProperty.length === 0 || !saveProperty) {
-            res.status(500).json({ status: 500, message: "Error in adding property", data: null });
-            console.log(error);
             return;
         }
         res.status(201).json({ status: 201, message: "Property added successfully", data: property });
     } catch (error) {
-        res.status(500).json({ status: 500, message: "Internal server error", data: null });
         console.log(error);
+        res.status(500).json({ status: 500, message: "Internal server error", data: null });
         return;
     }
 }
@@ -72,6 +72,8 @@ export const setProperty = async (req, res) => {
 export const getProperty = async (req, res) => {
     try {
         const userId = req.user?._id;
+
+        // check status to filter property data 
         const status = req.query?.status || req.body?.status;
 
         const limit = parseInt(req.body?.page_record) || 20;
@@ -86,8 +88,8 @@ export const getProperty = async (req, res) => {
             properties = dataprp;
         }
         else if (status) {
-            console.log("status->", status);
 
+            // joi validation for status
             const { error } = validateStatus({ status });
             if (error) {
                 res.status(400).json({ status: 400, message: error, data: null });
@@ -102,14 +104,15 @@ export const getProperty = async (req, res) => {
             properties = dataprp;
         }
 
+        // check property found or not
         if (properties?.length > 0) {
             const data = await Promise.all(properties.map(async (prop) => {
                 let propertyId = prop?._doc?._id;
-                // console.log("propertyId - >", propertyId);
 
+                // check milestone for property
                 const result = await Milestone.find({ propertyId: propertyId });
-                // console.log("result->", result);
 
+                // calculate paid and unpaid milestone
                 let paidMilestone = 0;
                 let unpaidMilestone = 0;
                 let totaldMilestone = result?.length || 0;
@@ -123,7 +126,6 @@ export const getProperty = async (req, res) => {
                         }
                     });
                 }
-
                 return { ...prop._doc, totaldMilestone, paidMilestone, unpaidMilestone };
             }));
             res.status(200).json({ status: 200, message: "Properties fetched successfully", data });
@@ -144,22 +146,55 @@ export const updateProperty = async (req, res) => {
         const propertyData = req.body;
         let propertyId = req.params?.id;
 
+        // check property id is valid or not
+        const isIDValid = isValidObjectId(propertyId);
+        if (!isIDValid) {
+            res.status(400).json({ status: "400", message: "Property id is not valid", data: null });
+            return;
+        }
+
+        // joi validation
         const { error } = validateProperty(propertyData);
         if (error) {
-            res.status(400).json({ status: 400, message: error, data: null });
             console.log(error);
+            res.status(400).json({ status: 400, message: error, data: null });
             return;
         }
 
-        if (!propertyId) {
-            res.status(400).json({ status: 400, message: "Property id is required", data: null });
-            return;
+        // check status is sold or purchase then check milestone paid or not
+        if (propertyData.status === "sold" || propertyData.status === "purchase") {
+            const result = await Milestone.find({ propertyId: propertyId });
+            if (result.length === 0) {
+                res.status(404).json({ status: 404, message: "Total Amount not Paid Yet, Milestone not found for this Property ", data: null });
+                return;
+            }
+
+            let paidAmount = 0;
+            let unpaidAmount = 0;
+            let totalAmount = parseInt(propertyData.total);
+
+            if (result.length > 0) {
+                result.forEach(obj => {
+                    if (obj.status === 'paid') {
+                        paidAmount += parseInt(obj.amount);
+                    } else {
+                        unpaidAmount += parseInt(obj.amount);
+                    }
+                });
+            }
+
+            // check total amount paid or not
+            if (paidAmount < totalAmount) {
+                res.status(400).json({ status: 400, message: "Total Amount not Paid Yet, Can not change the Status", data: null });
+                return;
+            }
         }
+
+        // exist property check
         let existproperty = await Property.findOne({ _id: { $ne: propertyId }, property_name: propertyData.property_name });
 
-
+        // check partner percentage should be 100
         let PartnerPercentage = checkPartnerPercentage(propertyData.partner);
-
         if (!PartnerPercentage) {
             console.log("Partner percentage should be 100");
             res.status(400).json({ status: 400, message: "Partner percentage should be 100", data: null });
@@ -167,6 +202,7 @@ export const updateProperty = async (req, res) => {
         }
 
         if (!existproperty) {
+            // update property data only with status not equal to sold
             const updatedProperty = await Property.findOneAndUpdate({ _id: propertyId, status: { $ne: "sold" } }, { ...propertyData }, { new: true });
             if (updatedProperty) {
                 res.status(200).json({ status: 200, message: "Property updated successfully", data: updatedProperty });
@@ -180,31 +216,38 @@ export const updateProperty = async (req, res) => {
         }
 
     } catch (error) {
-        res.status(500).json({ status: 500, message: "Internal server error", data: null });
         console.log(error);
+        res.status(500).json({ status: 500, message: "Internal server error", data: null });
         return;
     }
 }
- 
+
 export const deleteProperty = async (req, res) => {
     try {
         const propertyId = req.params?.id;
         const userId = req.user?._id;
 
+        // check property id is valid or not
+        const isIDValid = isValidObjectId(propertyId);
+        if (!isIDValid) {
+            res.status(400).json({ status: "400", message: "Property id is not valid", data: null });
+            return;
+        }
+
+        // find property and delete with status added only 
         const property = await Property.findOneAndDelete({ _id: propertyId, created_by: userId, status: "added" });
-        // console.log(property);
 
         if (property) {
+            // delete all milestone of property
             const result = await Milestone.deleteMany({ propertyId: propertyId, createdBy: userId });
-            // console.log("->",result);
             res.status(200).json({ status: 200, message: "Property deleted successfully", data: property });
         } else {
             res.status(404).json({ status: 404, message: "Can not delete Property with Status sold or deal_done or purchase ", data: null });
             return;
         }
     } catch (error) {
-        res.status(500).json({ status: 500, message: "Internal server error", data: null });
         console.log(error);
+        res.status(500).json({ status: 500, message: "Internal server error", data: null });
         return;
     }
 
@@ -213,13 +256,12 @@ export const deleteProperty = async (req, res) => {
 const formatDate = (date) => {
     const d = new Date(date);
     const day = String(d.getDate()).padStart(2, '0');
-    const month = String(d.getMonth() + 1).padStart(2, '0'); // Months are zero-based
+    const month = String(d.getMonth() + 1).padStart(2, '0');
     const year = d.getFullYear();
     return `${day}-${month}-${year}`;
 }
 
 const checkPartnerPercentage = (partnerData) => {
-    // console.log("partnerData->", partnerData);
 
     if (partnerData.length == 0 || partnerData === []) {
         return true;
@@ -229,8 +271,6 @@ const checkPartnerPercentage = (partnerData) => {
     partnerData.forEach(obj => {
         totalPercentage += parseInt(obj.percentage);
     });
-
-    console.log("totalPercentage->", totalPercentage);
 
     if (totalPercentage === 100) {
         return true;
@@ -275,10 +315,9 @@ const updateWallet = async (property, userId) => {
     }
     let newbalance = parseInt(property.total);
 
-    if(oldBalance-newbalance<=0){
+    if (oldBalance - newbalance <= 0) {
         return false;
     }
     const data = await Wallet.findOneAndUpdate({ userId: userId }, { balance: oldBalance - newbalance }, { new: true });
-
     return data;
 }
